requirements.txt
colorzero==2.0
gpiozero==2.0.1
hx711==1.1.2.3
joblib==1.5.2
numpy==2.3.5
pandas==2.3.3
python-dateutil==2.9.0.post0
pytz==2025.2
RPi.GPIO==0.7.1
scikit-learn==1.7.2
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.2
